wip; move instr to models/instrumentation
