+++ lib/cli/cli.js
@@ -5,10 +5,7 @@ var Promise       = require('../ext/promise');
 var getOptionArgs = require('../utilities/get-option-args');
 var logger        = require('heimdalljs-logger')('ember-cli:cli');
 var loggerTesting = require('heimdalljs-logger')('ember-cli:testing');
-var instrumentation = require('../utilities/instrumentation');
-var chalk = require('chalk');
-
-var instrumentationEnabled = instrumentation.instrumentationEnabled;
+var Instrumentation = require('../models/instrumentation');
 
 /**
  * @private
@@ -64,25 +61,12 @@ function CLI(options) {
 
   /**
    * @private
-   * @property initInstrumentation
+   * @property instrumentation
    */
-  this.initInstrumentation = options.initInstrumentation;
-
-  if (!options.initInstrumentation && instrumentationEnabled()) {
-    var heimdall = require('heimdalljs');
-    var initInstrumentationCookie = heimdall.start('init');
-    this.initInstrumentation = {
-      cookie: initInstrumentationCookie,
-      node: heimdall.current,
-    };
-
-    this.ui.writeLine(chalk.yellow(
-      'No init instrumentation passed to CLI.  Please update your global ember or ' +
-      'invoke ember via the local executable within node_modules.  Init ' +
-      'instrumentation will still be recorded, but some bootstraping will be ' +
-      'omitted.'
-    ));
-  }
+  this.instrumentation = new Instrumentation({
+    ui: options.ui,
+    initInstrumentation: options.initInstrumentation,
+  });
 
   logger.info('testing %o', !!this.testing);
 }
