+++ lib/models/builder.js
@@ -13,24 +13,9 @@ var attemptNeverIndex = require('../utilities/attempt-never-index');
 var findBuildFile = require('../utilities/find-build-file');
 var experiments = require('../experiments');
 var _resetTreeCache = require('./addon')._resetTreeCache;
-var instrumentation = require('../utilities/instrumentation');
-
-var vizEnabled = instrumentation.vizEnabled;
-var instrumentationEnabled = instrumentation.instrumentationEnabled;
 
 var Sync = require('tree-sync');
 
-function _enableFSMonitorIfInstrumentationEnabled() {
-  var monitor;
-  if (instrumentationEnabled()) {
-    var FSMonitor = require('heimdalljs-fs-monitor');
-    monitor = new FSMonitor();
-    monitor.start();
-  }
-  return monitor;
-}
-
-_enableFSMonitorIfInstrumentationEnabled();
 
 function _computeVizInfo(count, result, resultAnnotation) {
   var graph = require('heimdalljs-graph');
@@ -313,12 +298,12 @@ module.exports = CoreObject.extend({
       .then(function(result) {
         var vizJson;
 
-        if (instrumentationEnabled()) {
+        if (self.project.instrumentation.isEnabled()) {
           vizJson = self._computeVizInfo(self._buildCount, result, resultAnnotation);
           self._reportVizInfo(vizJson);
         }
 
-        if (vizEnabled()) {
+        if (self.project.instrumentation.isVizEnabled()) {
           vizJson.nodes = result.graph.__heimdall__.toJSONSubgraph();
           writeVizInfo(self._buildCount,{
             summary: vizJson.summary,
@@ -424,5 +409,4 @@ module.exports = CoreObject.extend({
   _computeVizInfo: _computeVizInfo,
 });
 
-module.exports._enableFSMonitorIfInstrumentationEnabled = _enableFSMonitorIfInstrumentationEnabled;
 module.exports._computeVizInfo = _computeVizInfo;
