+++ lib/models/project.js
@@ -50,6 +50,9 @@ function Project(root, pkg, ui, cli) {
     version: null,
     canNestRoots: false
   };
+
+  this.instrumentation = cli.instrumentation;
+  cli.instrumentation.project = this;
 }
 
 /**
